Time: 0 ms (100%), Space: 31.6 MB (13.08%) - LeetHub

+++ 0628-maximum-product-of-three-numbers/0628-maximum-product-of-three-numbers.cpp
@@ -27,7 +27,7 @@ public:
                 secondSmallest=nums[i];
             }
         }
-        int prod=max(largest*second*third, largest*smallest*secondSmallest);
+        long long prod=max(largest*second*third, largest*smallest*secondSmallest);
         return prod;
     }
 };
